[MA-8] Resolved issue: redefinition of qmetatypes. Metatypes are now inside their own file.
