+++ applications/mne_analyze/libs/anShared/Management/eventmanager.h
@@ -44,7 +44,7 @@
 
 #include "../anshared_global.h"
 #include "event.h"
-#include "../Utils/enums.h"
+#include "../Utils/types.h"
 
 
 //*************************************************************************************************************
