+++ applications/mne_analyze/libs/anShared/Model/abstractmodel.h
@@ -44,7 +44,7 @@
 //=============================================================================================================
 
 #include "../anshared_global.h"
-#include "../Utils/enums.h"
+#include "../Utils/types.h"
 
 
 //*************************************************************************************************************
