+++ applications/mne_analyze/libs/anShared/Model/surfacemodel.cpp
@@ -41,6 +41,7 @@
 //=============================================================================================================
 
 #include "surfacemodel.h"
+#include "../Utils/metatypes.h"
 #include "../Utils/types.h"
