+++ applications/mne_analyze/libs/anShared/Utils/types.h
@@ -66,13 +66,27 @@ using namespace Eigen;
 // DEFINE NAMESPACE MNEANALYZE
 //=============================================================================================================
 
-namespace MNEANALYZE
+namespace ANSHAREDLIB
 {
+    //=========================================================================================================
+    /**
+    * The MODEL_TYPE enum lists all available model types.
+    * Naming convention: NAMESPACE_CLASSNAME_MODEL
+    */
+    enum MODEL_TYPE
+    {
+        FSLIB_SURFACE_MODEL
+    };
 
+    //=========================================================================================================
+    /**
+    * Public enum for all available Event types.
+    */
+    enum EVENT_TYPE
+    {
+        PING,
+        DEFAULT
+    };
 } //NAMESPACE
-
-Q_DECLARE_METATYPE(Eigen::Vector3f);
-Q_DECLARE_METATYPE(Eigen::Vector3i);
-
 
 #endif // TYPES_H
